dts: tenderloin: use #define to switch i2c to uart
To help with testing, let's introduce a define that
can be switched from 0 to 1 to switch from I2C to
UART.

If I2C is defined, UART isn't redirected to serial1,
and the generic cyttsp driver is enabled on i2c.
If UART is defined, ttyMSM1 will be created, and the
cy8ctma395 driver will be loaded to discuss with ts_srv.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>